Objects Implementation Working
Seperated the collector and controller objects to seperate files and
debugged like crazy
