+++ Makefile.win
@@ -4,8 +4,8 @@
 CPP      = g++.exe
 CC       = gcc.exe
 WINDRES  = windres.exe
-OBJ      = main.o drive.o laptop.o adc_8591.o bluetooth.o brick.o color.o compass.o error.o filesystem.o i2c.o io_8574.o motor.o sensor.o sonar.o tilt.o armlift.o claw.o
-LINKOBJ  = main.o drive.o laptop.o adc_8591.o bluetooth.o brick.o color.o compass.o error.o filesystem.o i2c.o io_8574.o motor.o sensor.o sonar.o tilt.o armlift.o claw.o
+OBJ      = main.o drive.o laptop.o adc_8591.o bluetooth.o brick.o color.o compass.o error.o filesystem.o i2c.o io_8574.o motor.o sensor.o sonar.o tilt.o armlift.o claw.o datacollector.o controller.o
+LINKOBJ  = main.o drive.o laptop.o adc_8591.o bluetooth.o brick.o color.o compass.o error.o filesystem.o i2c.o io_8574.o motor.o sensor.o sonar.o tilt.o armlift.o claw.o datacollector.o controller.o
 LIBS     = -L"C:/Program Files (x86)/Dev-Cpp/MinGW64/lib" -L"C:/Program Files (x86)/Dev-Cpp/MinGW64/x86_64-w64-mingw32/lib" -static-libgcc lib/myo32.lib lib/myo64.lib
 INCS     = -I"C:/Program Files (x86)/Dev-Cpp/MinGW64/include" -I"C:/Program Files (x86)/Dev-Cpp/MinGW64/x86_64-w64-mingw32/include" -I"C:/Program Files (x86)/Dev-Cpp/MinGW64/lib/gcc/x86_64-w64-mingw32/4.9.2/include" -I"C:/Users/Matthew/Documents/GitHub/Gesture-Controlled-Robotic-Arm" -I"C:/Users/Matthew/Documents/GitHub/Gesture-Controlled-Robotic-Arm/myo"
 CXXINCS  = -I"C:/Program Files (x86)/Dev-Cpp/MinGW64/include" -I"C:/Program Files (x86)/Dev-Cpp/MinGW64/x86_64-w64-mingw32/include" -I"C:/Program Files (x86)/Dev-Cpp/MinGW64/lib/gcc/x86_64-w64-mingw32/4.9.2/include" -I"C:/Program Files (x86)/Dev-Cpp/MinGW64/lib/gcc/x86_64-w64-mingw32/4.9.2/include/c++" -I"C:/Users/Matthew/Documents/GitHub/Gesture-Controlled-Robotic-Arm" -I"C:/Users/Matthew/Documents/GitHub/Gesture-Controlled-Robotic-Arm/myo"
@@ -77,3 +77,9 @@ armlift.o: armlift.cpp
 
 claw.o: claw.cpp
 	$(CPP) -c claw.cpp -o claw.o $(CXXFLAGS)
+
+datacollector.o: datacollector.cpp
+	$(CPP) -c datacollector.cpp -o datacollector.o $(CXXFLAGS)
+
+controller.o: controller.cpp
+	$(CPP) -c controller.cpp -o controller.o $(CXXFLAGS)
