+++ Robotic Arm MYO.dev
@@ -29,7 +29,7 @@ IncludeVersionInfo=0
 SupportXPThemes=0
 CompilerSet=0
 CompilerSettings=0000000000000000000000000
-UnitCount=45
+UnitCount=49
 
 [VersionInfo]
 Major=1
@@ -494,6 +494,46 @@ BuildCmd=
 FileName=adc_8591.cpp
 CompileCpp=1
 Folder=Bluetooth Library
+Compile=1
+Link=1
+Priority=1000
+OverrideBuildCmd=0
+BuildCmd=
+
+[Unit47]
+FileName=datacollector.h
+CompileCpp=1
+Folder=
+Compile=1
+Link=1
+Priority=1000
+OverrideBuildCmd=0
+BuildCmd=
+
+[Unit46]
+FileName=datacollector.cpp
+CompileCpp=1
+Folder=
+Compile=1
+Link=1
+Priority=1000
+OverrideBuildCmd=0
+BuildCmd=
+
+[Unit49]
+FileName=controller.h
+CompileCpp=1
+Folder=
+Compile=1
+Link=1
+Priority=1000
+OverrideBuildCmd=0
+BuildCmd=
+
+[Unit48]
+FileName=controller.cpp
+CompileCpp=1
+Folder=
 Compile=1
 Link=1
 Priority=1000
