+++ Robotic Arm MYO.layout
@@ -1,10 +1,10 @@
 [Editors]
-Order=0,1,2,42,43,41,26,27,39,8,7
-Focused=0
+Order=0,1,2,42,43,41,26,27,39,8,7,45,46,44,47,48
+Focused=47
 [Editor_0]
-CursorCol=3
-CursorRow=310
-TopLine=298
+CursorCol=26
+CursorRow=55
+TopLine=26
 LeftChar=1
 [Editor_41]
 CursorCol=12
@@ -85,4 +85,24 @@ LeftChar=1
 CursorCol=3
 CursorRow=17
 TopLine=1
+LeftChar=1
+[Editor_45]
+CursorCol=2
+CursorRow=18
+TopLine=1
+LeftChar=1
+[Editor_46]
+CursorCol=25
+CursorRow=11
+TopLine=4
+LeftChar=1
+[Editor_47]
+CursorCol=21
+CursorRow=188
+TopLine=166
+LeftChar=1
+[Editor_48]
+CursorCol=7
+CursorRow=47
+TopLine=31
 LeftChar=1
