+++ main.cpp
@@ -9,6 +9,8 @@
 #include <algorithm>
 
 #include <myo.hpp>
+#include "controller.h"
+//#include "datacollector.h"
 
 #include <cstdlib>
 #include <conio.h>
@@ -37,380 +39,21 @@
 //ArmLift arm (motorLiftArm, elevatorTouch);
 //Claw claw (motorToggleClaw, motorRotateClaw);
 
-float motor_speed = 0;
-const int longSpeed[15]={75, 40, 25, 15, 10, 5, 0, 0, 0, -5, -10, -15, -25, -40, -75};
-const double latSpeed[9]={1,0.80,0.60,0,0,0,-0.60,-0.80,-1};
 
-
-
-bool controllingDrive, controllingArm, controllingClaw;
-
-void selectPort(ifstream &in, int &drivePort, int &armPort){
-	cout<<"\nSelect Connection Device"<<endl;
-	cout<<"Port"<<setw(20)<<"Name"<<endl;
-	
-	string n0;
-	int p0;
-	
-	while(in >> p0 >> n0){
-		cout<<setw(4)<<p0<<setw(20)<<n0<< endl;
-	}
-	
-	cout << "\nEnter the drive and arm com ports you would like to use " << endl;
-	cout << "from the above list or any other port you would like to use." << endl;
-	cin >> drivePort >> armPort;
-}
-
-class Controller{
-	private:
-	    Connection *driveConnection;
-	    Connection *armConnection;
-			
-	    Motor *motorLeftDrive;
-		Motor *motorRightDrive;
-		Motor *motorLiftArm;
-		
-		Motor *motorToggleClaw;
-		Motor *motorRotateClaw;
-		
-		Touch *elevatorTouch;
-		
-		Drive drive;
-		ArmLift arm;
-		Claw claw;
-		int rollInit,yawInit, target, index, lMotorSpeed,rMotorSpeed;
-		int longSpeed[15];
-		double latSpeed[9];
-		bool controllingDrive, controllingArm, controllingClaw
-  			 ;
-	
-	public:
-		
-	Controller(){
-		
-		*driveConnection = Bluetooth();
-		*armConnection = Bluetooth();
-
-		*motorLeftDrive = Motor (OUT_C, driveConnection);
-		*motorRightDrive = Motor (OUT_A, driveConnection);
-		
-		*motorLiftArm = Motor (OUT_B, driveConnection);
-		*motorToggleClaw = Motor (OUT_A, armConnection);
-		*motorRotateClaw = Motor (OUT_C, armConnection);
-		
-		*elevatorTouch = Touch(IN_1, driveConnection);
-		
-		drive= Drive(motorLeftDrive, motorRightDrive);
-		arm = ArmLift(motorLiftArm, elevatorTouch);
-		claw = Claw(motorToggleClaw, motorRotateClaw);
-		
-		double latSpeed[9]={1,0.80,0.60,0,0,0,-0.60,-0.80,-1};
-		int longSpeed[15]={75, 40, 25, 15, 10, 5, 0, 0, 0, -5, -10, -15, -25, -40, -75};
-		
-		target = 0;
-		yawInit = 0;
-		rollInit = 0;
-		index = 0;
-		lMotorSpeed = 0;
-		rMotorSpeed = 0;
-		controllingDrive = true;
-    	controllingArm = false;
-    	controllingClaw = false;
-
-	}
-	
-	bool setConnections (){
-		ifstream comIn ("comports.txt");
-		int driveCom = 0, armCom = 0;
-		selectPort(comIn, driveCom, armCom);
-		try{
-	
-			cout << "Trying to connect" << endl;
-	    	driveConnection->connect(driveCom);
-	    	armConnection->connect(armCom);
-	    	cout << "Connected" << endl;
-			arm.initialize();
-			claw.initialize();	
-		}
-		catch (const std::exception& e) {
-	        std::cerr << "Error: " << e.what() << std::endl;
-	        std::cerr << "Press enter to continue.";
-	        std::cin.ignore();
-	        drive.unlockDrive(false);
-	        arm.unlockArm(false);
-	        driveConnection->disconnect();
-	        armConnection->disconnect();
-	        return false;
-    		}
- 
-        myo::Hub hub("com.example.hello-myo");
-
-        std::cout << "Attempting to find a Myo..." << std::endl;
-
-        myo::Myo* myo = hub.waitForMyo(10000);
-
-    	if (!myo) {
-           throw std::runtime_error("Unable to find a Myo!");
-    	   }
-
-   	    std::cout << "Connected to a Myo armband!" << std::endl << std::endl;
-   	    return true;
-    }
-    
-    int getControlSystem(){
-    	
-    	if(controllingDrive){
-    		return 0;
-		}
-		if(controllingArm){
-			return 1;
-		}
-		return 2;
-							
-	}
-   
-    void setRoll(int roll){
-		if(onRollInit){
-			rollInit = roll;
-			onRollInit = false;
-		}
-		target = roll - rollInit;
-		if(target<=2){
-			claw.rotate(0);
-		} else if (target <= 11){
-			claw.rotate((target - 2) *12);
-		}
-	}
-	
-	void setPitch (int pitch){
-		if (controllingDrive && drive.isUnlocked()){
-			lMotorSpeed = rMotorSpeed = longSpeed[pitch_w];
-		} else if (controllingArm && arm.isUnlocked()){
-			arm.move(longSpeed[pitch_w]);
-		}
-	}
-	
-	void setYaw (int yaw){
-		if(onInit){
-			yawInit = yaw;
-			onInit=false;
-		}
-        index = yaw - yawInit + 4;
-        lMotorSpeed+=(100-fabs(lMotorSpeed))*(latSpeed[index]);
-        rMotorSpeed+=-(100-fabs(rMotorSpeed))*latSpeed[index];
-        
-        if(getControlSystem ==0){
-        	drive.forward(lMotorSpeed, rMotorSpeed);
-		}
-        
-	}
-
-   
-   	void runRobot(){
-   	char ch;
-   	
-   	DataCollector collector;
-	collectro.initialize(*this);
-    hub.addListener(&collector);
-
-    while (1) {
-        hub.run(1000/15);
-        collector.print();
-        if (kbhit()){
-        	ch = getch();
-			switch (ch){
-				case 32:
-					drive.unlockDrive(!drive.isUnlocked());
-					arm.unlockArm(!arm.isUnlocked());
-					claw.unlockClaw(!claw.isUnlocked());
-					std::cout << std::boolalpha << "Drive Control: " << drive.isUnlocked() << std::endl;
-					std::cout << std::boolalpha << "Arm Control: " << arm.isUnlocked() << std::endl;
-				 	break;
-				case 'a':
-					controllingArm = true;
-					controllingDrive = false;
-					controllingClaw = false;
-					std::cout << std::boolalpha << "Controlling Arm!" << std::endl;
-					break;
-				case 's':
-					controllingArm = false;
-					controllingDrive = false;
-					controllingClaw = true;
-					std::cout << std::boolalpha << "Controlling Claw!" << std::endl;
-				 	break;
-				case 'd':
-					controllingArm = false;
-					controllingDrive = true;
-					controllingClaw = false;
-					std::cout << std::boolalpha << "Controlling Drive!" << std::endl;
-				 	break;
-			}
-   	
-   }	
-
-	
-};
-class DataCollector : public myo::DeviceListener {
-private:
-		
-		Controller controller
-		int controlSystem;
-			
-public:
-    DataCollector()
-    : onArm(false), isUnlocked(false), roll_w(0), pitch_w(0), yaw_w(0), currentPose()
-    {
-    }
-    
-    void initialize(Controller *control){
-   		 controller = *control;
-   		 controlSystem = 0;
-   		 
-	}
-	
-    void onUnpair(myo::Myo* myo, uint64_t timestamp)
-    {
-        roll_w = 0;
-        pitch_w = 0;
-        index = 4;
-        yaw_w = 0;
-        onArm = false;
-        isUnlocked = false;
-    }
-	
-    void onOrientationData(myo::Myo* myo, uint64_t timestamp, const myo::Quaternion<float>& quat)
-    {
-        using std::atan2;
-        using std::asin;
-        using std::sqrt;
-        using std::max;
-        using std::min;
-		controlSystem = controller.getControlSystem();
-		if (controlSystem==2 && claw.isUnlocked()){
-			float roll = atan2(2.0f * (quat.w() * quat.x() + quat.y() * quat.z()),
-                           1.0f - 2.0f * (quat.x() * quat.x() + quat.y() * quat.y()));
-   			roll_w = static_cast<int>((roll + (float)M_PI/2.0f)/M_PI * 18);		   
-						   rollSpeed(roll);	
-		} else {
-			float pitch = asin(max(-1.0f, min(1.0f, 2.0f * (quat.w() * quat.y() - quat.z() * quat.x()))));
-			roll_w = static_cast<int>((roll + (float)M_PI/2.0f)/M_PI * 18);
-			controller.setPitch(pitch_w);
-		}
-        if (controlSystem == 0 && drive.isUnlocked()){
-			float yaw = atan2(2.0f * (quat.w() * quat.z() + quat.x() * quat.y()),
-                        1.0f - 2.0f * (quat.y() * quat.y() + quat.z() * quat.z()));
-			int yaw_w = static_cast<int>((yaw + (float)M_PI)/(M_PI * 2.0f) * 18);
-			controller.setYaw(yaw_w);
-			
-		}
-		
-        /*roll_w = static_cast<int>((roll + (float)M_PI)/(M_PI * 2.0f) * 18);
-       	pitch_w = static_cast<int>((pitch + (float)M_PI/2.0f)/M_PI * 18);
-        yaw_w = static_cast<int>((yaw + (float)M_PI)/(M_PI * 2.0f) * 18); 
-		index = yaw_w-yawInit+4;*/
-    }
-
-    void onPose(myo::Myo* myo, uint64_t timestamp, myo::Pose pose)
-    {
-        currentPose = pose;
-
-        if (pose != myo::Pose::unknown && pose != myo::Pose::rest) {
-            myo->unlock(myo::Myo::unlockHold);
-			
-			if (controllingClaw && claw.isUnlocked()){   
-				if(pose == myo::Pose::fingersSpread){
-					cout<<"Confirm Open Claw Request: "<<endl;
-					switch (getch()){
-						case 'y':
-							claw.open();
-							break;
-						default:
-							std::cout << "Cancelling open claw request." << endl;
-							break;	
-					}
-				} else if(pose == myo::Pose::fist){	
-					cout<<"Closing Claw";
-					claw.close();
-				}
-			}
-		
-			/*if (pose == myo::Pose::doubleTap){
-				myo->lock();
-			}*/
-
-            myo->notifyUserAction();
-        } else {
-            myo->unlock(myo::Myo::unlockHold);
-        }
-    }
-
-    void onArmSync(myo::Myo* myo, uint64_t timestamp, myo::Arm arm, myo::XDirection xDirection, float rotation,
-                   myo::WarmupState warmupState)
-    {
-        onArm = true;
-        whichArm = arm;
-    }
-
-    void onArmUnsync(myo::Myo* myo, uint64_t timestamp)
-    {
-        onArm = false;
-    }
-
-    void onUnlock(myo::Myo* myo, uint64_t timestamp)
-    {
-        isUnlocked = true;
-    }
-
-    void onLock(myo::Myo* myo, uint64_t timestamp)
-    {
-        isUnlocked = false;
-    }
-
-    void print()
-    {
-        string controllingNow;
-		std::cout << '\r';
-        if (controllingDrive){
-			controllingNow = "drive";
-		} else {
-			controllingNow = "arm"; 
-		}
-        std::cout   //<< '[' << std::string(roll_w, 'R') << std::string(18 - roll_w, ' ') << ']'
-//                  	<< '[' << std::string(pitch_w, 'P') << std::string(18 - pitch_w, ' ') << ']'
-//                  	<< '[' << std::string(yaw_w, 'Y') << std::string(18 - yaw_w, ' ') << ']'
-                  	<< '[' << controllingNow << index << ": " << latSpeed[index] << std::string(18 - index, ' ') << ']'
-                  	<< '[' << controllingNow << pitch_w << ": " << longSpeed [pitch_w] << std::string(18 - pitch_w, ' ') << ']';
-//                  << '[' << motorSpeed[pitch_w] << std::string(18 - pitch_w, ' ') << ']'
-//                  << '[' << motor_speed << std::string(18 - pitch_w, ' ') << ']';
-
-        if (onArm) {
-            std::string poseString = currentPose.toString();
-
-            std::cout << '[' << (isUnlocked ? "unlocked" : "locked  ") << ']'
-//                      << '[' << (whichArm == myo::armLeft ? "L" : "R") << ']'
-                      << '[' << poseString << std::string(14 - poseString.size(), ' ') << ']';
-        } else {
-//            std::cout << '[' << std::string(8, ' ') << ']' << "[?]" << '[' << std::string(14, ' ') << ']';
-        }
-
-        std::cout << std::flush;
-    }
-
-    bool onArm;
-    myo::Arm whichArm;
-
-    bool isUnlocked;
-
-    int roll_w, pitch_w, yaw_w;
-    myo::Pose currentPose;
-};
 
 
 
 int main(int argc, char** argv)
 {
+
 	Controller robotControl;
 	robotControl.setConnections();
-	
+	/*if(!robotControl.setConnections()){
+		cout << "FLAGGGG" << endl;
+		return -1;
+	}*/
+	cout << "FALGG" << endl;
+	robotControl.runRobot();
+	return 0;
 }
